Renamed UploadTasks into Upload + Some code/packages cleaning
