+++ STNetwork/src/commonMain/kotlin/com/infomaniak/multiplatform_swisstransfer/network/models/transfer/ContainerApi.kt
@@ -17,7 +17,7 @@
  */
 package com.infomaniak.multiplatform_swisstransfer.network.models.transfer
 
-import com.infomaniak.multiplatform_swisstransfer.common.interfaces.Container
+import com.infomaniak.multiplatform_swisstransfer.common.interfaces.transfers.Container
 import com.infomaniak.multiplatform_swisstransfer.network.serializers.DateToTimestampSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
